Move new personalization specs into spec directory.
