+++ test/unit/specs/components/Personalization/actions/setAttribute.spec.js
@@ -3,19 +3,19 @@ import {
   removeNode,
   appendNode,
   createNode
-} from "../../../../../src/utils/dom";
-import createMove from "../../../../../src/components/Personalization/actions/move";
+} from "../../../../../../src/utils/dom";
+import createSetAttribute from "../../../../../../src/components/Personalization/actions/setAttribute";
 
 const cleanUp = () => {
-  selectNodes("div#move").forEach(removeNode);
+  selectNodes("div#setAttribute").forEach(removeNode);
   selectNodes("style").forEach(node => {
-    if (node.textContent.indexOf("move") !== -1) {
+    if (node.textContent.indexOf("setAttribute") !== -1) {
       removeNode(node);
     }
   });
 };
 
-describe("Personalization::actions::move", () => {
+describe("Personalization::actions::setAttribute", () => {
   beforeEach(() => {
     cleanUp();
   });
@@ -26,22 +26,18 @@ describe("Personalization::actions::move", () => {
 
   it("should set personalized content", () => {
     const collect = jasmine.createSpy();
-    const move = createMove(collect);
+    const setAttribute = createSetAttribute(collect);
     const element = createNode("div", { id: "setAttribute" });
     const elements = [element];
 
     appendNode(document.body, element);
 
-    const settings = {
-      content: { left: "100px", top: "100px" },
-      meta: { a: 1 }
-    };
+    const settings = { content: { "data-test": "bar" }, meta: { a: 1 } };
     const event = { elements, prehidingSelector: "#setAttribute" };
 
-    move(settings, event);
+    setAttribute(settings, event);
 
-    expect(elements[0].style.left).toEqual("100px");
-    expect(elements[0].style.top).toEqual("100px");
+    expect(elements[0].getAttribute("data-test")).toEqual("bar");
     expect(collect).toHaveBeenCalledWith({
       meta: { personalization: { a: 1 } }
     });
